Restructure architecture: Move main.py to root and organize helpers into logical folders (#19)
* Initial plan

* Restructure architecture: move main.py to root and organize helpers


---------
