+++ helper/__init__.py
@@ -1,0 +1,4 @@
+"""
+Helper modules for KOMpass application.
+Contains organized submodules for authentication, configuration, processing, UI, and utilities.
+"""
