+++ helper/auth/__init__.py
@@ -1,0 +1,3 @@
+"""
+Authentication module for Strava OAuth integration.
+"""
