+++ helper/ui/ui_components.py
@@ -11,12 +11,12 @@ import time
 from streamlit_folium import st_folium
 import hashlib
 
-from route_processor import RouteProcessor
-from weather_analyzer import WeatherAnalyzer
-from auth_manager import get_auth_manager
-from config import get_config
-from logging_config import get_logger, log_function_entry, log_function_exit, log_error, log_execution_time
-from units import UnitConverter
+from ..processing.route_processor import RouteProcessor
+from ..processing.weather_analyzer import WeatherAnalyzer
+from ..auth.auth_manager import get_auth_manager
+from ..config.config import get_config
+from ..config.logging_config import get_logger, log_function_entry, log_function_exit, log_error, log_execution_time
+from ..utils.units import UnitConverter
 
 logger = get_logger(__name__)
